Show all entries works
The ability to show all entries currently works. The menu for showing entries has also been updated to show the name and allow the user to skip ahead to a cerain entry number. The ability to search entries and the ability to show all entries by a ceratin user still needs to be added.